requirements.txt
# 核心依赖
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# 数据处理
beautifulsoup4>=4.12.0
pdfplumber>=0.10.0
python-docx>=1.1.0
PyPDF2>=3.0.0

# 向量数据库
chromadb>=0.4.18
# requirements.txt 添加
tenacity>=8.2.0
sentence-transformers>=2.2.2  # 备用本地向量化

# 阿里云DashScope SDK
dashscope>=1.14.0
openai>=1.3.0  # 用于兼容OpenAI格式的调用

# 其他
requests>=2.31.0
PyYAML>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.0
loguru>=0.7.0

# 中文NLP
jieba>=0.42.1

# 可视化
plotly>=5.18.0